requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
fastapi-mail
python-dotenv
bcrypt==4.0.1
pytest
httpx